# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jupiter Token Health Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  details?: string | null;
}

export interface TokenSearchResult {
  address: string;
  name: string;
  symbol: string;
  /** @nullable */
  logoURI?: string | null;
  verified: boolean;
  decimals: number;
  tags?: string[];
  /** @nullable */
  organicScore?: number | null;
  /** @nullable */
  daily_volume?: number | null;
}

/**
 * Raw response from Jupiter API
 */
export type TokenSearchResponseRawResponse = { [key: string]: unknown };

export interface TokenSearchResponse {
  results: TokenSearchResult[];
  /** Raw response from Jupiter API */
  rawResponse?: TokenSearchResponseRawResponse;
}

/**
 * Raw response from Jupiter API
 */
export type TokenMetadataRawResponse = { [key: string]: unknown };

export interface TokenMetadata {
  address: string;
  name: string;
  symbol: string;
  /** @nullable */
  logoURI?: string | null;
  verified: boolean;
  decimals: number;
  tags?: string[];
  /** @nullable */
  organicScore?: number | null;
  /** @nullable */
  organicScoreLabel?: string | null;
  /** @nullable */
  daily_volume?: number | null;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  market_cap?: number | null;
  /** @nullable */
  fdv?: number | null;
  /** @nullable */
  holder_count?: number | null;
  /** @nullable */
  liquidity?: number | null;
  /** Raw response from Jupiter API */
  rawResponse: TokenMetadataRawResponse;
}

/**
 * Extra price info if available
 */
export type TokenPriceDataExtraInfo = { [key: string]: unknown };

export interface TokenPriceData {
  id: string;
  type?: string;
  price: string;
  /** Extra price info if available */
  extraInfo?: TokenPriceDataExtraInfo;
}

/**
 * Raw response from Jupiter API
 */
export type TokenPriceResponseRawResponse = { [key: string]: unknown };

export interface TokenPriceResponse {
  mintAddress: string;
  /** @nullable */
  price?: string | null;
  priceData?: TokenPriceData;
  /** @nullable */
  timeTaken?: number | null;
  /** Raw response from Jupiter API */
  rawResponse: TokenPriceResponseRawResponse;
}

export interface WalletTokenHolding {
  mint: string;
  amount: string;
  uiAmount: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  symbol?: string | null;
  /** @nullable */
  logoURI?: string | null;
  /** @nullable */
  verified?: boolean | null;
  /** @nullable */
  usdPrice?: number | null;
  /** @nullable */
  usdValue?: number | null;
  /** @nullable */
  organicScore?: number | null;
  /** @nullable */
  decimals?: number | null;
  isFrozen: boolean;
}

export type WalletPortfolioResponseRawBalances = { [key: string]: unknown };

export interface WalletPortfolioResponse {
  walletAddress: string;
  /** @nullable */
  totalUsdValue?: number | null;
  tokenCount: number;
  holdings: WalletTokenHolding[];
  rawBalances: WalletPortfolioResponseRawBalances;
}

export type SearchTokensParams = {
  /**
   * Token name, symbol, or mint address to search for
   */
  query: string;
};