# lib/api-zod/src/generated/types/walletPortfolioResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jupiter Token Health Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletPortfolioResponseRawBalances } from "./walletPortfolioResponseRawBalances";
import type { WalletTokenHolding } from "./walletTokenHolding";

export interface WalletPortfolioResponse {
  walletAddress: string;
  /** @nullable */
  totalUsdValue?: number | null;
  tokenCount: number;
  holdings: WalletTokenHolding[];
  rawBalances: WalletPortfolioResponseRawBalances;
}