# lib/api-zod/src/generated/types/walletTokenHolding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jupiter Token Health Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletTokenHolding {
  mint: string;
  amount: string;
  uiAmount: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  symbol?: string | null;
  /** @nullable */
  logoURI?: string | null;
  /** @nullable */
  verified?: boolean | null;
  /** @nullable */
  usdPrice?: number | null;
  /** @nullable */
  usdValue?: number | null;
  /** @nullable */
  organicScore?: number | null;
  /** @nullable */
  decimals?: number | null;
  isFrozen: boolean;
}